pyproject.toml
[project]
name = "pdf-ocr-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
]
